# FWC.RMS.WebApi/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Revenue Management System",
    "description" : "Florida Fish and Wildlife Conservation Commision - Revenue Management System.",
    "termsOfService" : "http://swagger.io/terms/",
    "contact" : {
      "email" : "apiteam@fwc.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
...
  "servers" : [ {
    "url" : "https://fws.swagger.io/v1"
  }, {
    "url" : "http://fws.swagger.io/v1"
  } ],
...
  "paths" : {
...
    "/transmittals" : {
      "get" : {
        "tags" : [ "transmittal" ],
        "summary" : "get transmittal records",
        "operationId" : "getTransmittals",
        "parameters" : [ {
          "name" : "transmittalStatus",
          "in" : "query",
          "description" : "Transmittal Status",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Transmittal"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          },
          "500" : {
            "description" : "Internal server error",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          }
        }
      },
      "post" : {
        "tags" : [ "transmittal" ],
        "summary" : "Create a Transmittal record",
        "operationId" : "createTransmittals",
        "requestBody" : {
          "description" : "Transmittal request object",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTransmittalRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Transmittal"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          },
          "500" : {
            "description" : "Internal server error",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          }
        }
      }
    },
    "/transmittals/{transmittalNumber}" : {
      "patch" : {
        "tags" : [ "transmittal" ],
        "summary" : "Update a Transmittal record",
        "operationId" : "updateTransmittals",
        "parameters" : [ {
          "name" : "transmittalNumber",
          "in" : "path",
          "description" : "Transmittal Number",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "description" : "Transmittal request object",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTransmittalRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Transmittal"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          },
          "500" : {
            "description" : "Internal server error",
            "content" : { },
            "x-codegen-request-body-name" : "body"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CreateTransmittalRequest" : {
        "required" : [ "transmittalDate" ],
        "type" : "object",
        "properties" : {
          "transmittalDate" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "UpdateTransmittalRequest" : {
        "type" : "object",
        "properties" : {
          "transmittalTotalCount" : {
            "type" : "integer"
          },
          "transmittalTotal" : {
            "type" : "number",
            "format" : "double"
          },
          "transmittalStatus" : {
            "type" : "string"
          }
        }
      },
      "Transmittal" : {
        "required" : [ "transmittalDate", "transmittalNumber", "transmittalStatus" ],
        "type" : "object",
        "properties" : {
          "transmittalNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "transmittalDate" : {
            "type" : "string",
            "format" : "date"
          },
          "transmittalTotalCount" : {
            "type" : "integer"
          },
          "transmittalTotal" : {
            "type" : "number",
            "format" : "double"
          },
          "transmittalStatus" : {
            "type" : "string"
          }
        },
        "example" : {
          "transmittalTotalCount" : 6,
          "transmittalTotal" : 1.4658129805029452,
          "transmittalStatus" : "transmittalStatus",
          "transmittalNumber" : 0,
          "transmittalDate" : "2000-01-23"
        }
      },
...
    },
    "securitySchemes" : {
      "fws_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "http://fws.swagger.io/oauth/dialog",
            "scopes" : {
              "write:transmittals" : "modify transmittals in your account",
              "read:transmittals" : "read your transmittals"
            }
          }
        }
      },
      "api_key" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      }
    }
  }
}